+++ src/OpenTelemetry.Instrumentation.GrpcNetClient/CHANGELOG.md
@@ -6,6 +6,9 @@
   telemetry collection to be filtered on a per-request basis.
   ([#5236](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5236))
 
+* Updated OpenTelemetry core component version(s) to `1.19.1`.
+  ([#5329](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5329))
+
 ## 1.19.0-beta.1
 
 Released 2026-Sep-18